+++ packages/content/src/ContentStoreService.ts
@@ -15,10 +15,10 @@
  */
 
 import { Observable } from 'rxjs';
-
 import { crafterConf, SDKService } from '@craftercms/classes';
 import { CrafterConfig, Descriptor, Item } from '@craftercms/models';
 import { composeUrl } from '@craftercms/utils';
+import { map } from 'rxjs/operators';
 
 /**
  * Returns an Item from the content store.
@@ -32,7 +32,7 @@ export function getItem(path: string, config?: CrafterConfig): Observable<Item> 
   return SDKService.httpGet(requestURL, { url: path, crafterSite: config.site }, config.headers);
 }
 
-export interface GetDescriptorConfig {
+export interface GetDescriptorConfig extends CrafterConfig {
   flatten: boolean;
 }
 
@@ -42,15 +42,25 @@ export interface GetDescriptorConfig {
  * @param {CrafterConfig & GetDescriptorConfig} config? - The config override options to use
  */
 export function getDescriptor(path: string): Observable<Descriptor>;
-export function getDescriptor(path: string, config: Partial<CrafterConfig & GetDescriptorConfig>): Observable<Descriptor>;
-export function getDescriptor(path: string, config?: Partial<CrafterConfig & GetDescriptorConfig>): Observable<Descriptor> {
-  config = crafterConf.mix(config);
-  const requestURL = composeUrl(config, config.endpoints.GET_DESCRIPTOR);
-  return SDKService.httpGet(requestURL, {
+export function getDescriptor(path: string, config: Partial<GetDescriptorConfig>): Observable<Descriptor>;
+export function getDescriptor(path: string, config?: Partial<GetDescriptorConfig>): Observable<Descriptor> {
+  let cfg = crafterConf.mix(config);
+  return SDKService.httpGet<Descriptor>(composeUrl(cfg, cfg.endpoints.GET_DESCRIPTOR), {
     url: path,
-    crafterSite: config.site,
-    flatten: Boolean(config.flatten)
-  }, config.headers);
+    crafterSite: cfg.site,
+    flatten: Boolean(config?.flatten)
+  }, cfg.headers).pipe(
+    // Manually introduce the path into the response as descriptor endpoint does not return it.
+    map((descriptor: Descriptor) => {
+      let prop = typeof descriptor.page === 'undefined' ? 'component' : 'page';
+      return {
+        [prop]: {
+          ...descriptor[prop],
+          localId: path
+        }
+      };
+    })
+  );
 }
 
 /**