+++ packages/content/src/utils.ts
@@ -14,24 +14,11 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-import { ContentInstance, Descriptor, Item } from '@craftercms/models';
-
-interface GraphQLResponse {
-  [root: string]: {
-    // Repeating groups & node selectors collection root
-    item?:
-      object[] |
-      Array<{
-        key: string;
-        value: string;
-        component: unknown;
-      }>
-    // Other field types
-    [contentTypeField: string]: unknown
-  }
-}
-
-type DescriptorResponse = Descriptor | Item | GraphQLResponse;
+import { ContentInstance, DescriptorResponse } from '@craftercms/models';
+import { urlTransform } from './UrlTransformationService';
+import { getDescriptor, GetDescriptorConfig } from './ContentStoreService';
+import { map, switchMap } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 
 const systemPropMap = {
   guid: 'id',
@@ -65,13 +52,20 @@ const systemProps = Object.keys(systemPropMap).concat(
   Object.values(systemPropMap)
 );
 
-export function parseDescriptor(data: DescriptorResponse | DescriptorResponse[]): ContentInstance | ContentInstance[] {
+export interface ParseDescriptorOptions {
+  /** Whether to parse the field values to their respective data type based on postfix (e.g. _i number, _b boolean, etc) */
+  parseFieldValueTypes: boolean;
+}
+
+export function parseDescriptor(data: DescriptorResponse, options?: ParseDescriptorOptions): ContentInstance;
+export function parseDescriptor(data: DescriptorResponse[], options?: ParseDescriptorOptions): ContentInstance[];
+export function parseDescriptor(data: DescriptorResponse | DescriptorResponse[], options?: ParseDescriptorOptions): ContentInstance | ContentInstance[] {
   if (data == null) {
     return null;
   } else if (Array.isArray(data)) {
-    return data.map((item) => parseDescriptor(item) as ContentInstance);
+    return data.map((item) => parseDescriptor(item, options) as ContentInstance);
   } else if (data.children) {
-    return parseDescriptor(extractChildren(data.children));
+    return parseDescriptor(extractChildren(data.children), options);
   } else if (data.descriptorDom === null && data.descriptorUrl) {
     // This path catches calls to getChildren (/api/1/site/content_store/children.json?url=&crafterSite=)
     // The getChildren call contains certain items that can't be parsed into content items.
@@ -94,10 +88,10 @@ export function parseDescriptor(data: DescriptorResponse | DescriptorResponse[])
       sourceMap: {}
     }
   };
-  return parseProps(extractContent(data), parsed);
+  return parseProps(extractContent(data), parsed, options);
 }
 
-export function parseProps<Props = object, Target = object>(props: Props, parsed: Target = {} as Target): Target {
+export function parseProps<Props = object, Target = object>(props: Props, parsed: Target = {} as Target, options: ParseDescriptorOptions = { parseFieldValueTypes: false }): Target {
   Object.entries(props).forEach(([prop, value]) => {
     if (ignoreProps.includes(prop)) {
       return; // continue, skip prop.
@@ -130,7 +124,7 @@ export function parseProps<Props = object, Target = object>(props: Props, parsed
       }
       parsed[prop] = parsed[prop].map((item) => {
         const { key, value, component, include } = item;
-        if ((item.component) || (item.key && item.value)) {
+        if ((item.component) || (item.key && item.include)) {
           // Components
           const newComponent = {
             label: value,
@@ -143,17 +137,59 @@ export function parseProps<Props = object, Target = object>(props: Props, parsed
                   : component?.path ? component.path : null
               )
           };
-          return parseDescriptor(newComponent);
+          return parseDescriptor(newComponent, options);
         } else {
-          // Repeat group items
-          return parseProps(item);
+          // Repeat group items or key/value pairs
+          return parseProps(item, void 0, options);
         }
       });
     } else {
-      parsed[prop] = value ?? null;
+      parsed[prop] = value != null
+        ? options.parseFieldValueTypes ? parseFieldValue(prop, value) : value
+        : null;
     }
   });
   return parsed;
+}
+
+export function parseFieldValue(propName: string, propValue: any): number | string | boolean {
+  let postFix = propName.substr(propName.lastIndexOf('_'));
+  switch (postFix) {
+    /*
+    _i  For integer number.
+    _l  For long integer number.
+    _f  For floating point number.
+    _d  For long floating point number.
+    _to For time
+    _dt For date in ISO 8601
+    */
+    case '_b':
+      return propValue.toLowerCase().trim() === 'true';
+    case '_i':
+    case '_l':
+    case '_f':
+    case '_d':
+      return parseFloat(propValue);
+    default:
+      return propValue;
+  }
+}
+
+export function fetchModelByPath(path: string): Observable<ContentInstance>;
+export function fetchModelByPath(path: string, options?: GetDescriptorConfig & ParseDescriptorOptions): Observable<ContentInstance>;
+export function fetchModelByPath(path: string, options?: GetDescriptorConfig & ParseDescriptorOptions): Observable<ContentInstance> {
+  return getDescriptor(path, { flatten: options?.flatten ?? true }).pipe(
+    map((descriptor) => parseDescriptor(descriptor, { parseFieldValueTypes: options?.parseFieldValueTypes ?? true }))
+  );
+}
+
+export function fetchModelByUrl(webUrl: string): Observable<ContentInstance>;
+export function fetchModelByUrl(webUrl: string, options?: GetDescriptorConfig & ParseDescriptorOptions): Observable<ContentInstance>;
+export function fetchModelByUrl(webUrl: string, options?: GetDescriptorConfig & ParseDescriptorOptions): Observable<ContentInstance> {
+  return urlTransform('renderUrlToStoreUrl', webUrl).pipe(
+    switchMap((path) => getDescriptor(path as string, { flatten: options?.flatten ?? true })),
+    map((descriptor) => parseDescriptor(descriptor, { parseFieldValueTypes: options?.parseFieldValueTypes ?? true }))
+  );
 }
 
 /**