+++ packages/models/src/GraphQLResponse.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2022 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License version 3
@@ -14,13 +14,17 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-import { Item } from './item';
-import { GraphQLResponse } from './GraphQLResponse';
-
-export interface Descriptor {
-  page?: any;
-  component?: any;
-  [prop: string]: any;
+export interface GraphQLResponse {
+  [root: string]: {
+    // Repeating groups & node selectors collection root
+    item?:
+      object[] |
+      Array<{
+        key: string;
+        value: string;
+        component: unknown;
+      }>
+    // Other field types
+    [contentTypeField: string]: unknown
+  }
 }
-
-export type DescriptorResponse = Descriptor | Item | GraphQLResponse;